[elastic-agent-system-metrics] Add GetOneRootEvent method to process handler (#150)
## What does this PR do?

This is part of https://github.com/elastic/elastic-agent/issues/4083 

This adds a `GetOneRootEvent()`. method to `ProcStats` we need this for
https://github.com/elastic/elastic-agent/issues/4083 , as that issue
requires metricbeat's system/process metricset to monitor a specific
PID, which it can't do now. This will have to be followed up with a
change to beat itself to rope this in. This also cleans up a bit of code
and adds some docs.

## Why is it important?

needed for https://github.com/elastic/elastic-agent/issues/4083 

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added an entry in `CHANGELOG.md`