Remove hardcoded module definitions in CI (#39506)
For reasons described in [^1], we hard coded `aws` or `kubernetes` for the env var `MODULE`, when no file changes were detected under `modules/` per project.

This commit reverts this and allows all module tests to run if no changesets apply. As discussed in [^1], for the tests to run properly (and not go into an infinite loop)
if there are no module changes the `MODULE` env var should **not** be set, rather than set to an empty string.

This commit is an important fix as it covers a major delta in test coverage between Jenkins and Buildkite.

Closes https://github.com/elastic/ingest-dev/issues/2993

[^1]: https://github.com/elastic/ingest-dev/issues/2993#issuecomment-2104741977

(cherry picked from commit c91503b4cb33b4d98befb1dcd9bdbd888d44d28b)