Remove certain steps from running for Branches
This commit makes some adjustments in the step conditions after
reviewing the Jenkins files and how we perform the step validations in
Jenkinsfile.

If there is a key `when` inside the Jenkinsfile step it will validate the required
conditions and if there is no `branches: true` then it will not include that
step.

Signed-off-by: Alexandros Sapranidis <alexandros@elastic.co>