[add_session_metadata processor] Populate events with user/group names
Enrich process events with user and group names for the process, and related
processes (parent, session leader, group leader, entry leader).

A new cache for user and group names is added to the process DB, so that the
names only need to be read once, and not on every process event that's received.