[BK] - Remove certain steps from running for Branches (#39533)
This commit makes some adjustments in the step conditions after
reviewing the Jenkins files and how we perform the step validations in
Jenkinsfile.

If there is a key `when` inside the Jenkinsfile step, it will validate the required
conditions, and if there are no `branches: true`, then it will not include that
step.

Signed-off-by: Alexandros Sapranidis <alexandros@elastic.co>
(cherry picked from commit 176e9b8c6dc55ef211f39f194d1366b593793691)