Remove shipper (#39584)
This commit removes all code related to the shipper and fixes lint warnings.

The deprecated `ioutil.ReadDir` has been replaced by `io.ReadDir`.
A single for loop is used to iterate over the dirEntries and if there is an
error reading the FileInfo from a DirEntry, we log it and continue to the next
dirEntry. This effectively adds more resilience to the original code
that would not process any segment if the call to `ioutil.ReadDir` had
failed.