Log raw events to a separate log file (#38767)
This commit introduces a new logger core that can be configured through logging.event_data and is used to log any message that contains the whole event or could contain any sensitive data. This is accomplished by adding log.type: event to the log entry. The logger core is responsible for filtering the log entries and directing them to the correct files.

At the moment it is used by multiple outputs to log indexing errors containing the whole event and errors returned by Elasticsearch that can potentially contain the whole event.

The debug processor is also using the new logger core. The "Publish event:..." log entries are now directed to the event log file.

---------

Co-authored-by: Pierre HILBERT <pierre.hilbert@elastic.co>