heartbeat - move seccomp customization to InitializeModule
Move the overwriting of the arm64 seccomp policy out of the Go
init() phase and into a function call that is specific to Heartbeat. This
will prevent Heartbeat's policy from affecting all other Beats that run
as part of Agentbeat.

Note that other customizations from all Beats that use seccomp.ModifyDefaultPolicy
within init() are being combined in Agentbeat to create a policy that is the aggregation of
all seccomp customizations. So the allow-list type policies are likely more permissive than
expected.

Fixes #39754