Add migrate_checkpoint config option
The new migrate_checkpoint config option controls if the input v2
should perform a migration check on start.

If migrate_checkpoint is true, the input checks and performs the
migration (if v1 info exists) on the very first v2 run.

If migrate_checkpoint is false, the input will skip the migration
assistant and will not perform any checks or migration.