[libbeat] Remove "producer cancel" features from queue API (#39760)
"Producer cancel" is a feature that allows closing queue producers to also cancel any pending events created by that producer that have not yet been sent to a queue reader. It was introduced as a small part of a [very large refactor](https://github.com/elastic/beats/pull/4492) in 2017, but current code doesn't depend on it for anything. Since this feature adds considerable complexity to the queue API and implementation, this PR removes the feature and associated helpers.

This PR should cause no user-visible behavior change.