[elastic-agent-libs] Fix closerCore With behaviour (#210)
This commit fixes the behaviour of `closerCore.With` and add tests to
ensure it works as expected.

The tests are a modification of existing tests that calls the
configuration functions like Beats does and ensures both closerCore
and typedCore are used. The method calls get proxyed from one core to
another, so both are tested together.

The race detector is enabled when testing on Linux and MacOS. Windows
requires CGO and installation of gcc on CI, so it is not covered by this commit.