Fail integration tests if requirements are not met
If the system running integration tests doesn't have the requirements, fail
the test.

If the tests are incorrectly marked as passing, it's less likely that anyone
will notice that no tests were actually run, and will cause misleading results,
and not catch errors may exist.

It's better to fail on systems that don't have the requirements, so such systems
can be investigated and fixed.