packetbeat/route: fix test failure on Windows 2022 (#39822)
Windows 2022 appears to have added new text to the output of netsh. The text
is interpreted as having three fields. This impact on the behaviour of the
field splitting which expects five.

- Originally this would cause a panic due to bounds checking. This was fixed
  in #39757 by erroring on an unexpectedly short line.
- The fieldsN helper expects lines to be n fields or longer, but does not
  terminate on lines with fewer fields.

So treat the first blank line found in the table as the end of the table and
fix the fieldsN helper.

(cherry picked from commit 35f8d09c79361c4367dada1fa1de0d1ab0704819)