[libbeat] Add a metrics observer to the queue (#39774)
Add a metrics observer to the queue, reporting the metrics:

- `queue.added.{events, bytes}`, the number of events/bytes added to the queue
- `queue.consumed.{events, bytes}`, the number of events/bytes sent to the outputs
- `queue.removed.{events, bytes}`, the number of events/bytes removed from the queue after acknowledgment (`queue.removed.events` is an alias for the existing `queue.acked`).
`queue.filled.{events, bytes}`, the current number of events and bytes in the queue (gauges)

It also fixes the behavior of `queue.filled.pct.events`, renaming it `queue.filled.pct`.

All byte values reported by the memory queue are 0 if the output doesn't support early encoding.

This required some refactoring to the pipeline, which previously used a single custom callback to track its only queue metric (`queue.acked`) from `outputObserver`, and also used that to manage a wait group that was used to drain the queue on pipeline shutdown. The main changes are:
- A new interface type, `queue.Observer`, with an implementation `queueObserver` for standard metrics reporting.
- `queueMaxEvents` and `queueACKed` were removed from `pipeline.outputObserver`, since their logic is now handled by `queue.Observer`.
- A queue factory now takes a `queue.Observer` instead of an ACK callback
- The queue API now includes a `Done()` channel that signals when all events are acked / shutdown is complete, so shutdown handling now waits on that channel in `outputController.Close` instead of the shared waitgroup in `Pipeline.Close`.
- `pipeline.outputObserver` was renamed `pipeline.retryObserver` since its only remaining functions track retries and retry failures. It is now owned by `eventConsumer` (its only caller) instead of `pipeline.outputController`.

The queue previously had a `Metrics()` call that was used in the shipper but didn't integrate with Beats metrics. It had no remaining callers, so I deleted it while adding the new helpers.