refactor: drop golang.org/x/tools/go/vcs dependency
golang.org/x/tools/go/vcs is deprecated and was initially developed before module
support was added.
However we are not using vcs.Cmd at all, we're using one method to retrieve the
vcs being used. As a quick fix, we can copy the method since it's short and
self-contained and we don't need the whole package.