x-pack/filebeat/input/{cel,salesforce}: check response for nilness before logging (#40144)
It turns out that retryablehttp will pass a nil *http.Response to the
client's ErrorHandler function although this is not documented. In the
cases where we are using this, this will result in a nil deference
panic when the retries exceed their maximum. So, check for nilness to
avoid this.

(cherry picked from commit b056515c51dcd043fc33b512eab31cc6d631e325)