fix broken since timestamp
The since tests was broken because newer versions of Jouranlctl
accepts time stamps on RFC3339, but older versions do not.

The Reader now formats the since parameter in a format accepted by
older versions of systemd. The since tests ensures the timestamp is in
the local timezone to avoid conflicts with jouranlct.