WIP - Load Pipeline object lazily
Delay instantiation of Pipeline object till the first client (input)
tries to connect to it.
The idea is to use a builder object for the Pipeline on which we can set
parameters (in this case the global processor config) until the Pipeline
is instantiated.
Once the Pipeline is instantiated, the global processor reloader will
return a specific error, signaling that to reload configuration we need
a new Pipeline (i.e. a beat restart).
At this point in the code the reload and restart seem to work, however
it seems that the output unit is not set correctly (we probably need
some additional hook to correctly wire the Pipeline and the Output
object