filebeat/decode_cef - Add option to ignore empty values (#40268)
Added option to ignore empty values in the decode_cef processor.

In the decode_cef processor, when there are empty values in the extensions section, we get errors during log parsing. This change provides a flag in decode_cef config to override this default behavior and ignore the fields with empty value. Some example errors that this helps handle are:

    error in field 'cn1': strconv.ParseInt: parsing "": invalid syntax
    error in field 'destinationTranslatedAddress': value is not a valid IP address

Closes #40236