not log start up info when on managed mode
When on managed mode the beats no longer log its startup information: the build info, the host info, the Go info and the process info.
The host info is the same collected by the Elastic Agent and all the others are known or can be found based on the Elastic Agent version.