Manage AWS SDK(s) version with Dependabot (#39738)
Set up Dependabot to manage the AWS SDK version.

With the current reactive and manual process, our dependencies are often outdated. To release a bugfix to a dependency, we need to wait for the following stack release instead of merging it shortly after it's available from AWS.

See https://github.com/elastic/beats/issues/39492 to learn more.