x-pack/filebeat/i.../entitya.../.../okta: Avoid a negative request rate (#40267)
This is the minimal change necessary to fix the following problem.

Around the time of a rate limit reset, if current time is after the
reset time returned in response headers, the rate limiting code will
set a negative target rate, and if that's done at a time when no
request budget has accumulated, it will not recover and will wait
forever.

(cherry picked from commit 5087dd45412f872414e30a3aeddacbfed5fd6782)