feat: bump mongodb library to drop aws v1 indirect dep
we are using a 3y old version that causes the aws sdk v1 to be
included as an indirect dependency.
Bump to current stable to get rid of it.
The libmongocrypt C library is required when using client-side encryption.
Specific versions of libmongocrypt are required for different
versions of the Go Driver:
Go Driver v1.5.0 requires libmongocrypt v1.1.0 or higher.
Go Driver v1.12.0 requires libmongocrypt v1.8.0 or higher.