x-pack/filebeat/input/entityanalytics/provider/azuread: fix configuration ordering (#40487)
Previously, the configuration was eager, this caused a panic when the request
trace logging option was turned on since it was started without a context.

Move the construction of components to when we have all the parts we need.

(cherry picked from commit dd73639b1a3c0dd96e568bbb50c98cc26d4732c9)