x-pack/filebeat/input/entityanalytics/provider/azuread/fetcher/graph: set log max_size (#40494)
There was no check on the tracer log settings, so the max_size was not being
checked for zero. This would result in 100MB log rotation trigger and zero-sized
bodies. There is no validation logic to hook this into in the config for the
fetcher, so calculate it immediately before use.