[elastic-agent-system-metrics] fix non-fatal error tests (#176)
## What does this PR do?

So, a previous PR changed the behavior of the process code such that
we're now returning non-fatal errors.
However, a bunch of the tests were not updated, and are now failing. 

I'm a little worried by the fact that none of the tests failed
initially, as they fail when I run them locally. Part of the problem is
that the CI environment often doesn't have any processes running as
other users, meaning we won't hit a lot of code that deals with
permissions errors. Still, it feels like _something_ should have failed
here. Still investigating.

## Why is it important?

This is causing tests here to fail:
https://github.com/elastic/beats/pull/40507

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added an entry in `CHANGELOG.md`