tests: fix check to enable the race detector in supported platforms
This code is almost 4 years old, the support for the race detector was
lacking at that point and Elastic had its first steps with arm64 builds of Beats.

I updated the check to allow tests to run with RACE_DETECTOR=true in supported
platforms, this includes linux/arm64 and darwin/arm64.