Fix elasticsearch re-connection after network error
When the Elasticsearch client fails to publish events, it ends up
calling `Close` in the connection (that is reused). To cancel the
in-flight requests, the context is cancelled and a new one is created
to used in future requests.

The callback to check the version holds a reference to the connection
via a closure, now the Elasticsearch client holds a pointer to that
connection, so whenever Close is called, the callback can create a
request with the new, not cancelled, context.

An integration test is added to ensure the
ES output can always recover from network errors.