x-pack/filebeat/input/internal/httplog: log headers instead of request/response (#41072)
Add logging of structured fields for request and response headers and drop
logging of request/response using httputil. We lose no significant
information from dropping the request logging; in requests there will be
an "Accept-Encoding: gzip" that we do not get in the headers, but this
is always present in all requests from beats clients. The benefit of
including headers as structured data is that we gain the ability to more
easily search logs for header values.