[filebeat][GCS] - Fixed failed job handling and removed false-positive error logs (#41142)
[filebeat][GCS] fix failed job handling and remove false-positive error logs

In cases where GCS report an error, we were logging the error but not
otherwise making use of the information. When GCS was reporting that a
requested object does note exist this was causing unnecessary log-spam and
not updating the set of failed jobs to remove the object, resulting in future
futile re-attempts to collect the object. When any other error was reported,
the set of failed jobs was not being updated, resulting in continued
re-attempts, even past the maximum retry count.

This fixes both cases by differentiating the two situations, logging only at
debug level when the object is reported to be missing and removing the object
from the failed rework set, and logging at error and increasing the failure
count for all other error cases.

(cherry picked from commit fbff3189de92dcede7fb524e2957d1c51433f1f2)