Correctly set 'input.type' when using the system integration (#41246)
The system integration now uses the new input system-logs that is just
a proxy to instantiate either the log input or journald. That caused
the 'input.type' in the final event to contain 'system-logs' instead
of 'log' or 'journald'. This PR fixes it buy adding a set step in the
ingest pipeline.

The ingest pipeline is used because Beats sets this filed outside of
the input code.

Co-authored-by: Pierre HILBERT <pierre.hilbert@elastic.co>