Update parsers test to use ndjson parser
This commit updates the parse test to use ndjson parser instead of
multiline because the multiline parser can have issues when journald
input is reading from files. There is a corner case where the
journalctl exits successfully and the reader goroutine gets an error,
this makes Next to return early, making the multiline to also return
early.

So far I have only seen this happening when reading from file and at
the very end of the file, hence it does not seem to be a critical bug.