x-pack/filebeat/input/entityanalytics/provider/activedirectory: improve modification time and deletion logic (#41179)
This improves the update time stamps of modified events by using the
documents' whenChanged fields in the case of returned documents, and the
current time when a document is identified as having been deleted. The latest
of these is used to determine the time filter for the next Active Directory
query.

Documents are marked as deleted only when they are found to not exist in full
sync collection, and are removed from the state store when they are
identified as deleted. The change in behaviour to not use updates to identify
corrects behaviour that would cause older but not deleted entities to be
deleted from the index.

(cherry picked from commit 6b540745cb3c990b9320bf2f1380bb1b4579a31f)