Read journal entries from all boots (#41244)
Some versions of journalctl will only return messages from the current
boot when --follow is passed, it will even ignore the cursor or date
arguments.

This commit reads messages from all boots by first calling journalctl
without the --follow flag, reading all entries and once it
successfully exits, then we restart journalctl with the cursor and the
--follow flag.

The parser test is upsted to use ndjson parser instead of
multiline because the multiline parser can have issues when journald
input is reading from files. There is a corner case where the
journalctl exits successfully and the reader goroutine gets an error,
this makes Next to return early, making the multiline to also return
early.

TestJournaldInput assumed journalctl was run only once, however that
has changed recently. The test is updated to accommodate for that and rename
to TestJournaldInputRunsAndRecoversFromJournalctlFailures, which
better reflects what it is actually testing.

---------

Co-authored-by: Pierre HILBERT <pierre.hilbert@elastic.co>