x-pack/filebeat/input/{cel,httpjson}: fix flaky test (#41358)
Do not use uncontrolled randomisation in tests where it is not necessary. The
retry tests in the httpjson and cel packages were using a randomised 5xx HTTP
status code to trigger the retry behaviour of the the go-retryablehttp
package. This has the unfortunate consequence of causing 2% of test runs to
fail.

The reason for this is given in the go-retryablehttp documentation[1]:

	Mainly, if an error is returned by the client (connection errors,
	etc.), or if a 500-range response code is received (except 501),
	then a retry is invoked after a wait period.

Since the package is already tested, and is documented to accept all 5xx
status codes except 501 to cause a retry, just use 500.

[1]https://pkg.go.dev/github.com/hashicorp/go-retryablehttp@v0.7.7#section-readme