Lower logging levels in add_session_metadata processor
Reduce logging levels for some log messages in the add_session_metadata processor.

If something goes wrong with enrichment, many of these logs would be called for every process event, resulting
in a lot of logging spam. The logs that have been changed to `Debug` are logs that could potentially be called on
every enrichment, so they are lowered to below the default log level.

The other logs that have been changed to `Warn` or `Info` will only be called once, or are related to a timer, so
they will not cause a large amount of logs to be created.

There are better ways to detect if enrichment has failed, so changing the log levels shouldn't negatively affect anything.
For example an Elasticsearch query or alert on missing data that only this processor will populate (i.e.
`process.entry_leader` fields) will show processes that weren't properly enriched.