[add_session_metadata] Always use correct code for backend in use.
The config backend option and actual backend is use doesn't always match; the
'auto' option doesn't match a real backend (kernel_tracing, procfs). This fixes
some logic so that when the 'auto' option is used, the processor will always follow
the code path intended for whatever the actual backend is use is.