Lower logging levels in add_session_metadata processor (#41409)
Reduce logging levels for some log messages in the add_session_metadata processor.

If something goes wrong with the process cache where all, or most, processes are missed, many of these logs would be called for every process event, resulting in a lot of logging spam. These logs have been changed to Debug, which is below the default log level and will not cause log spam.

The logs that have been reduce to Info are in a timer, so they will not cause a lot of spam, but they should be Informational messages

There are better ways to detect if enrichment has failed, so changing the log levels shouldn't negatively affect anything. For example, an Elasticsearch query or alert on missing fields that only this processor will populate will show processes that weren't properly enriched (i.e. process.entry_leader fields).