Update go-quark to v0.2.0 (#41408)
Update go-quark to v0.2.0. This version of go-quark includes a bugfix from libquark that fixes use of kprobes
on newer Linux kernels.

This also removes the license override for go-quark; the license in it should now be detected properly, removing the need for the override.

(cherry picked from commit f8d3880d13194ae9aaafcfaddcda21c7845cdf14)