Revert system module support for journald (#41489)
Revert the system module usage of the system-logs input that was
enabling it to run the Journald input. The revert is done in the
system module configuration, pipelines and documentation. The
system-logs input and its tests are kept.

(cherry picked from commit 00d7161a1d70e9bb422bfbfefd4122700472d8be)

# Conflicts:
#	filebeat/module/system/auth/config/auth.yml
#	filebeat/module/system/syslog/config/syslog.yml
#	filebeat/tests/integration/systemlogs_linux_test.go