otelconsumer: handle entity too large errors
The elasticsearchexporter does not handle a 413 Request Entity too Large from
Elasticsearch, only forwarding the error back to the client.

When using the batcher config in the ES exporter, it runs synchronously, any
error reported can be intercepted in the otelconsumer. When using the batch
processor this happens asynchronously and there is no way to handle the
error. If we can intercept an entity too large error, split the batch and
retry.