go.mod - use elastic/goja fork
Move to the elastic/goja fork instead of andrewkroh/goja.
This the same code just moved into the elastic org and tagged
with v2019-01-28+beats.

For background, this fork contains a small change to goja that
allows arbitrary map types (such as elastic's "MapStr" type) to
be treated as javascript objects without requiring any conversion.

Also note that since the time of the fork in 2019, there is now
a path to accomplish this without a fork by implementing a
`toValue(r *goja.Runtime) goja.Value` method on MapStr. It would
mean introducing a direct dependency on goja. A tested example of
can be found in the linked gist.

https://gist.github.com/andrewkroh/19eccc0844bb935222914234c9510aa1

https://github.com/elastic/goja/releases/tag/v2019-01-28%2Bbeats