Put full version of Go in go.mod while updating
The previous configuration was always replacing the full version of Go
with a minor version: 1.22.8 -> 1.22.

Newer versions of Go now put the full version value in `go.mod`. This leads
to our CI failing because `go mod tidy` causes a change in `go.mod`.

This is to fix those failures.